requirements.txt
# FastAPI and server dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
email-validator==2.0.0.post2

# Firebase integration
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-auth==2.22.0

# Vector database
chromadb
# pydantic==2.3.0

# Document processing
PyPDF2==3.0.1
pdfminer.six==20221105
python-pptx==0.6.21
jupytext==1.15.1
nbformat==5.9.2
markdown==3.4.4
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3

# NLP and embedding
openai==1.3.0
tiktoken==0.5.1
numpy==1.26.4
nltk==3.8.1
spacy==3.7.4

# HTTP client
aiohttp==3.8.5
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
tqdm==4.66.1
loguru==0.7.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1

# Development tools
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1